+++ CHANGELOG.md
@@ -1,2 +1,4 @@
+* Added fromAny and tryFromAny
+
 ## v0.1.0 (2024-06-14)
 * Built initial implementation